requirements.txt
ansi2html==1.8.0
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
Flask==2.3.2
gunicorn==20.1.0
httpie==3.2.1
idna==3.4
install==1.3.5
ipykernel==6.23.1
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter-dash==0.4.2
jupyter_client==8.2.0
jupyter_core==5.3.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
plotly==5.14.1
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.7.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
requests==2.30.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.3.5
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tenacity==8.2.2
tornado==6.3.2
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
Werkzeug==2.3.4
whitenoise==6.4.0
